* obsolete G1UpdateBufferSize
G1UpdateBufferSize has previously been used to size the refinement
buffers and impose a minimum limit on the number of cards per thread
that need to be pending before refinement starts.

The former function is now obsolete with the removal of the dirty
card queues, the latter functionality has been taken over by the new
diagnostic option `G1PerThreadPendingCardThreshold`.

I prefer to make this a diagnostic option is better than a product option
because it is something that is only necessary for some test cases to
produce some otherwise unwanted behavior (continuous refinement).

CSR is pending.