8352147
Hi all,

  please review this refactor of the TestEagerReclaimHumongousRegionsClearMarkBits test that runs way too long, and sometimes even causing timeouts in GHA.

So the problem is that TestEagerReclaimHumongousRegionsClearMarkBits checks whether after eager reclaim during marking the mark on the humongous object is cleared correctly. It does so with a trial-and-error approach allocating humongous objects and hoping that the faulty state somehow occurs.

This can take a long time, and although the test limits itself to 50s runtime, for some reason there can still be sporadic timeouts in some setups.

However for a long time it is possible to halt concurrent mark just before completion, inducing the exact state needed for this test. So rewrite the test to be more targeted.

Testing: gha, running test locally

Thanks,
  Thomas

* also check for actual region reclamation
* last minute typo