8353950: Ensure JDK does not crash when running reproducer
For investigation/testing issue JDK-8353950 it is necessary to create a
non-crashing baseline. Testing JDK-8353950 was done with a fastdebug
build of the JDK. The build was modified so that instead of blockingly
showing a messagebox, only CLI output was generated when an assertion
was hit. This is necessary, so that the reproducer runs with higher
concurrency.

The call to MessageBoxA was commented out for this in awt_Debug.cpp.

Reproducer Code
===============

```java

import java.awt.Toolkit;
import java.awt.datatransfer.FlavorEvent;
import java.awt.datatransfer.FlavorListener;
import java.awt.datatransfer.StringSelection;
import java.util.concurrent.atomic.AtomicBoolean;
import javax.swing.SwingUtilities;

public class Clipboardreproducer2 {

    private static final System.Logger LOG = System.getLogger(Clipboardreproducer2.class.getName());

    public static void main(String[] args) throws Exception {

        AtomicBoolean end = new AtomicBoolean(false);

        SwingUtilities.invokeLater(() -> {
            Toolkit.getDefaultToolkit().getSystemClipboard().addFlavorListener(new FlavorListener() {
                @Override
                public void flavorsChanged(FlavorEvent e) {
                    LOG.log(System.Logger.Level.INFO, "Flavor Changed: {0} // {1}", e, SwingUtilities.isEventDispatchThread());
                }
            });
        });

        Runnable copy = new Runnable() {
            @Override
            public void run() {
                try {
                    Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new StringSelection("Test"), null);
//                    LOG.log(System.Logger.Level.INFO, "Copy 1");
                } catch (IllegalStateException ex) {
                    LOG.log(System.Logger.Level.ERROR, "Failed to invoke copy (1)", ex);
                }
                if(! end.get()) {
                    SwingUtilities.invokeLater(this);
                } else {
                    System.out.println("Exitting");
                }
            }
        };

        SwingUtilities.invokeLater(copy);

        Thread.sleep(60 * 1000);

        LOG.log(System.Logger.Level.INFO, "DONE");

        end.set(true);

        Thread.sleep(500);

        LOG.log(System.Logger.Level.INFO, "END");

        System.exit(0);

    }

}
```

Crash message
=============

```
#
# A fatal error has been detected by the Java Runtime Environment:
#
#  Internal Error (c:\temp\jdk\src\hotspot\share\runtime/jniHandles.inline.hpp:81), pid=3816, tid=5224
#  assert(external_guard || result != nullptr) failed: Invalid JNI handle
#
# JRE version: OpenJDK Runtime Environment (25.0) (fastdebug build 25-internal-adhoc.Matthias.jdk)
# Java VM: OpenJDK 64-Bit Server VM (fastdebug 25-internal-adhoc.Matthias.jdk, mixed mode, tiered, compressed oops, compressed class ptrs, g1 gc, windows-amd64)
# Problematic frame:
# V  [jvm.dll+0x5160eb]  JNIHandles::resolve_impl<0,0>+0x27b
#
# Core dump will be written. Default location: c:\Temp\hs_err_pid3816.mdmp
#
# An error report file with more information is saved as:
# c:\Temp\hs_err_pid3816.log
[16.331s][warning][os] Loading hsdis library failed
#
# If you would like to submit a bug report, please visit:
#   https://bugreport.java.com/bugreport/crash.jsp
#
```

The code from JDK-8332271 was also tested, but that did not yield a
visible result. It crashes, but the only data recorded is an entry in
the eventlog of windows.

Analysis
========

Looking through the code in `awt_Clipboard.cpp` there is a possible race
condition. Each call to openClipboard with a non-null caller will reset
the value of AwtClipboard::theCurrentClipboard. That variable holds a
reference to the java peer of the native clipboard.

The race is (line number refer to state in commit
1d7138fe021d64686ac59cddd1bc7c787c610aa9):

Thread 1: Application code calls Clipboard#setContents. On windows the
code now flows through WClipboard#openClipboard and will run to
awt_Clipboard.cpp line 145 (this is in the native implementatio of
openClipboard. Just before DeleteGlobalRef.

Thread 2: The windows event loop gets a WM_CLIPBOARDUPDATE message and
calls into AwtClipboard::WmClipboardUpdate (awt_Clipboard.cpp line 60).
The code runs through the non-null check and just before the
`CallVoidMethod` call.

Thread 1: Continues and executes DeleteGlobalRef.

Thread 2: Continues exection with:
`env->CallVoidMethod(theCurrentClipboard, lostSelectionOwnershipMID)`.
But at this point theCurrentClipboard does not hold a valid object
handle anymore and crashes the VM.

Proposed solution
=================

The code already makes assumptions, that there is only one AwtToolkit
(see calls to AwtToolkit::GetInstance). In addition
WToolkit#getSystemClipboard ensures, that only one WClipboard will
be created for this.

Based on this, this change drops the ability to pass in a new Clipboard
on each call to `openClipboard`. Instead the registration of the
java side of the clipboard with the native side is moved to construction
time of the WClipboard object. `openClipboard` is modified, so that it
verifies the assumption asserting that a new owner matches the initial
owner.