8353950: Ensure windows clipboard is not accessed concurrently
The focus of this change is the windows port of the AWT. openClipboard
and closeClipboard are only implemented there.

The general idea here is, that the section between openClipboard and
closeClipboard must be considered a critical section, that must not be
accessed concurrently.

The fix idea here is to use a ReentrantLock to protect the native side.
openClipboard now check if the lock can be locked (tryLock) if that
fails, an IllegalStateException is raised.

This is defined behavior for the accessors from client code (see for
example the documentation for Clipboard#setContents and
Clipboard#getContents.

The remaining problem is the upcall from the WToolkit. The message loop
handler has to react to the WM_CLIPBOARDUPDATE. When that message is
received an upcall to WClipboard#handleContentsChanged is invoked. That
in turn will make a downcall to the native getClipboardFormats. That
method used EnumClipboardFormats to enumerate the formats in the
clipboard. The problem is, that EnumClipboardFormats need the clipboard
to be opened (as implemented in openClipboard). 

This is problematic as the upcall from native should not lead to an
exception. Even if the exception is caught, it will result in flavor
listeners not being called.

Since Windows Vista / Windows Server 2008 GetUpdatedClipboardFormats
can be used, which does not need an open clipboard and thus would be
safe to be called from the upcall.

Testing:

- fastdebug build does not hit asserts anymore for the reproducer code
  (see previous commit)
- reproducer code for JDK-8332271 does not crash anymore. Code was
  run for multiple minutes and no crash was observed
- tests from test/jdk/java/awt/Clipboard were run and tests succeeded
  in release and fastdebug configuration