Adjustments to heuristics
Previous configuration appears to be stable.  But we're still observing
suboptimal behavior.  In particular, on the Extremem medium workload, we
see three back-to-back old collections, where each is delayed until the
previous old has finished its mixed evacuations and the
coalesce-and-fill efforts that follow it.  Independently, a study of
specjbb 2015 suggests that G1 and ZGC are both much more relaxed about
how much they allow Old-gen to expand without any effort to collect it.
And this works well for their specjbb scores.  So we want to be a bit
more lazy about old collections.

  ShenandoahIgnoreOldGrowthBelowPercentage defaults to 40 (not 10)
  ShenandoahDoNotIgnoreGrowthAfterYoungCycles defaults to 500 (not 50)
  ShenandoahOldGrowthPercent defaults to 25 (not 12.5)

On the same Extremem medium workload, we observe that when we are trying
"desperately" to make progress on both old-gen and young-gen collections
concurrently, the old is being starved and not making quick enough
progress because it only gets 80 ms time slices between young cycles and
it is not surging its own efforts.  So I want to allow old to surge more
quickly (not attenuate its surges), and I think it best to allow young
to run more quickly when it is under duress (by not attenuating its
surges).  The intention is that if young can run more quickly, there
will be an increase of idle time between young cycles, during which old
can get more than its minimal time slice.

The hope is that the combination of allowing the important and urgent
old collection to start earlier (because it doesn't have to wait for
a previous unnecessary and redundant old collection to finish) and
more aggressive surging of workers once we start old will allow old
to finish in a more timely and effective manner.