Gentle refinements to heuristics after debugging
Be less eager to promote by shared allocation.  Note that failure to
promote is "not" harmless because we have not reserved for evacuation of
these objects in young reserve.  But it is slightly easier (usually) to
redirect an evacuation to an existing mutator region than to flip an
empty mutator region.