Increase EvacWaste defaults
A behavior manifest in this development branch is that there is greater
variation in collection set evacuation workloads.  When we happen to
have very large evacuation workloads, more of the evacuation effort is
handled by mutator threads, causing the available reserves to be
scattered between a larger number of GCLABs and PLABs.  The end results
is that we are much more likely to experience OOM during evac.

Plan to restore original default values when we implement a new
OOM-during-evac protocol that does not require upgrade to Full GC.