8356320: fix tpix when we flood-fill a rect and tpix changes index
In this ticket the problem was:
`tpix` was the transparent pixel index of the current `model`. However `decoder.saved_image` uses its own IndexColorModel (see `decoder.saved_model`).

In a simple gif these two models are the same, but in this case the gif has some frames that change the transparent pixel index between frames. We were filling the wrong tpix value, resulting in an opaque rectangle.