8365047: Remove exception handler stub code in C2
The C2 exception handler stub code is only a trampoline to the
generated exception handler blob. This change removes the extra
step on the way to the generated blob.

According to some comments in the source code, the exception handler
stub code used to be patched upon deoptimization, however presumably
these comments are outdated as the patching upon deoptimization happens
for post-call NOPs only.