8356061: restructuring to move things to EDT
This was requested here:
https://github.com/openjdk/jdk/pull/25244#discussion_r2091656355

mrserb asked: "Also please move creation and access_to all Swing components to EDT"

(I'm not sure this will help much? My understanding was *creation* of Swing components could happen off the EDT as long as they were made displayable on the EDT.)

Now we still call jc.getLocationOnScreen off the EDT. If that posed a thread-based problem it'd probably manifest as a IllegalComponentStateException, which is not mentioned in 8356061.