8354646: Make sure AquaTextPasswordFieldUI can't use selectWordAction
The core problem here was we could call selectWordAction. This problem was already solved in the BasicPasswordFieldUI, but the AquaTextPasswordFieldUI (and SynthPasswordFieldUI) do NOT extend the BasicPasswordFieldUI.

So the problem is partially about multiple inheritance. This is my first response at solving the problem: move everything to the parent BasicTextUI class and use an instanceof to make sure it is only applied to JPasswordFields.