8334759
Hi all,

  please review this change that "fixes" the `TestMixedGCLiveThreshold.java` test by requiring the thread factory used for these tests to not use virtual threads instead of platform threads.

That adds some additional memory consumption, and since the test is about testing G1 reaction due to particular known memory consumption, it can fail.

The fix is to add the appropriate `@requires` tag that has been introduced in jtreg 7.5.1 (which is current default for builds, https://bugs.openjdk.org/browse/JDK-8334759).

Testing: verified that the test is not run if the virtual thread test thread factory is used.

Thanks,
  Thomas