8361582: AArch64: Some ConH values cannot be replicated with SVE
After this commit -
https://github.com/openjdk/jdk/commit/a49ecb26c5ff2f949851937f3bb036d7946a103e,
the JTREG test -
hotspot/jtreg/compiler/vectorization/TestFloat16VectorOperations.java
fails for some of the tests which contain constant values such as -

public void vectorAddConstInputFloat16() {
         for (int i = 0; i < LEN; ++i) {
             output[i] = float16ToRawShortBits(add(shortBitsToFloat16(input1[i]), FP16_CONST));
         }
     }
<The full failure log is present in the JBS ticket, thus not
reproducing it here>

The current code in the JDK results in the generation of sve_dup
instruction for every 16-bit immediate while the acceptable range is
[-128, 127] for 8-bit immediates and [-127 << 8, 128 << 8] with a
multiple of 256 for 16-bit signed immediates.

This patch allows the generation of sve_dup instruction for only those
16-bit values which are within the limits as specified above and for the
values which are out of range, the immediate half float value is loaded
from the constant pool into a register ("loadConH" mach node) which is
then replicated or broadcasted to an SVE register ("replicateHF" mach
node).