8365147: AArch64: Replace DMB + LD + DMB with LDAR for C1 volatile loads
Replaces the DMB ISH + LD + DMB ISHLD sequence with LDAR
for volatile field loads - for example, AtomicLong::get.

This is valid, as originally the DMBs were necessary due to
the case described here - https://bugs.openjdk.org/browse/JDK-8179954.
As in the rare case where the LD can be reordered with an LDAR
or STLR from the C2 implementation for stores and loads, these
DMBs are required.
However, acquire/release operations use a sequentially consistent model
which does not allow reordering between them. Hence, the LD can be
replaced with an LDAR to disallow reordering with a STLR/LDAR
and the first DMB can be removed.

The LDAR has acquire semantics, so it's impossible for
memory accesses after to be reordered before; the DMB ISHLD is
not required. Therefore, a singular LDAR is sufficient.

This excludes floats and doubles, as they do not have
equivalent load-acquire instructions.

Change-Id: Ia93607f8bb20c2d974fe6b2e586dd3239bb2728c