SimpleClassTypeSignature: intern member names used for generic type signatures
The MethodRepository constructs SimpleClassTypeSignature instances for each
method with generic signatures used by reflection. These are constructed once
and held as long as the method is. Since many methods refer to the same class
name, this name copy ends up potentially in the heap many thousands of times.

Full context:
https://mail.openjdk.org/pipermail/core-libs-dev/2025-August/150022.html