8365296: Build failure with Clang due to -Wformat warning after JDK-8364611
This issue is related to the definition of __sigset_t (__sigset_t  act.sa_mask).  
In the glibc source, __sigset_t is defined in multiple places:  
- bits/types/__sigset_t.h  
- sysdeps/unix/sysv/linux/bits/types/__sigset_t.h (introduced after glibc 2.25)  
During compilation with Clang, the latter definition appears to be used, where __sigset_t is a struct, causing the printf("%X", act.sa_mask) to fail.