8303904: designing around imagined edge cases
I don't know if these will ever come up in the real world, but since this is critical infrastructure I want to be careful. (A Swing app may become unusable if this code fails.)

Concern A:
In the event our AffineTransform is a rotation or a flip:
Now we'll use the c.createImage(virtualWidth, virtualHeight), which means we fall back to the code we've been using for over a decade.

Concern B:
In the event our AffineTransform scales to zero, we'll at least make the image 1x1.