8365656
Hi all,

  please review this change that fixes a div-by-zero in calculating liveness for logging.

It is possible that the young gen group cardset is empty, making the liveness calculation divide by zero.

The change special-cases this case when printing - since the young gen card set group is kind of special and always exists, also print it always. An alternative would be not printing it at all in this case. Feel free to argue for that solution. Also the printed value for liveness (0.0%) is up for discussion.

Testing: failing test case does not fail any more now and then, gha

Thanks,
  Thomas