8373696: AArch64: C1: Embed FP constants into the main code
The constants section for C1 on AArch64 is only used for fp32/fp64
constants. Embed the values into the instruction section like for
the integers.

Use immediates where it doesn't require more memory than would have
been with the constants section. This is always the case for fp32.
Otherwise, use LDR + B skip_label + <int64>.

This removes the constants section - saving memory which otherwise
would be used for padding to align the entry point to CodeEntryAlignment.

Also, this enables further potential optimizations that require size of
the constants section to be known before code is emitted.