8368205: [TESTBUG] VectorMaskCompareNotTest.java crashes when MaxVectorSize=8
The VectorShape size of `I_SPECIES_FOR_CAST` declared in test **VectorMaskCompareNotTest.java**
is half that of `L_SPECIES_FOR_CAST`. And `L_SPECIES_FOR_CAST` is created with the maximum shape.
Therefore, if `MaxVectorSize` is set to 8, the shape size of `I_SPECIES_FOR_CAST` is 4, which
is an illegal value because the minimum vector size requirement is 8 bytes.

This pull request addresses the issue by ensuring that this test runs only when `MaxVectorSize`
 is set to 16 bytes or higher.