8369818
Hi all,

  please review this update to the `gc+heap=debug` output to contain all region types (I do not know why it only showed eden and survivor).

This is also printed in other places (like `hs_err` files) and it is sometimes very interesting to know that e.g. there are humongous regions around.

It also makes the unit MB because KB is too unwieldy, and the minimum region size is 1 MB anyway.

Testing: local compilation, checking the log output

Thanks,
  Thomas