Explicitly hoist marking context and index
In performance critical loops that implement freeset rebuild and choose
collection set, it is common to call ShenandoahHeapRegion::get_live_data_words()
and functions that depend on its implementation including:
get_live_data_bytes(), garbage(), has_live(), and
ShenandoahCollectionSet::add_region().

The refactored implementation calculates ShenandoahMarkingContext at
loop prologue, and remembers region index as an induction variable
within the loop.

This saves multiple indirections in the implementations of these
methods.

With these changes, performance of traditional Shenandoah on specjbb2015
is restored to on par, or slighly better than before the change to
fix-live-data-for-mixed-evacation-candidates.

(The implementation of fix-live-data-for-mixed-evacuation-candidates
makes it slightly more expensive to calculate live-data-words.  This
resulted in a performance regression on specjbb2015 before this commit.)