Detect cyclic includes with Files::isSameFile
checkCyclicInclude() is invoked after we successfully get an InputStream
for the path to avoid skipping the same IOException several times inside
checkCyclicInclude() if the path doesn't exist.

Also, perform symlinks resolution only in the following cases:
 • When we need to resolve a relative include
 • For clarity to the user in logging messages
 • For clarity to the user in exception messages

In the first case, the resolution is a requirement, in the last two
cases it is a nice-to-have. But given the last two are exceptional
cases anyway, we let any resolution error bubble up.