Move Hotspot Agent debugger for MacOSX
The Mac OSX implementation of the Serviceability Agent and related code
is quite different from what's needed from the BSD implementation. Still
we have tried to keep the coexisting in one codebase in the out-of-tree
BSD port, as that's where the OSX code has been living.

This sometimes cause problems when updates to the Mac OSX port breaks
the BSD implementation.

As we are working on getting the BSD port into a state for future
upstreaming to the mainline repo, this patch clears the path by moving
the Mac OSX implementation of the Servicability Agent to a more fitting
namespace.

This should allow us to proceed with the BSD implementation undisturbed,
but also without risking breaking the OSX port.

This work was sponsored by The FreeBSD Foundation